requirements.txt
alpaca-py>=0.43.0
python-dotenv>=1.0.0
pandas>=2.0.0
yfinance>=0.2.0
